mad2/requirements.txt
flask-security-too
Flask-Cors
flask-restful
flask 
Flask-SQLAlchemy
setuptools
argon2_cffi
celery
flask-caching
redis
Flask-Excel